+++ src/soundio.c
@@ -586,7 +586,7 @@ int soundio_outstream_set_volume(struct SoundIoOutStream *outstream, double volu
     struct SoundIo *soundio = outstream->device->soundio;
     struct SoundIoPrivate *si = (struct SoundIoPrivate *)soundio;
     struct SoundIoOutStreamPrivate *os = (struct SoundIoOutStreamPrivate *)outstream;
-    return si->outstream_set_volume(si, os, volume);
+	return (int)(si->outstream_set_volume(si, os, (float)volume));
 }
 
 static void default_instream_error_callback(struct SoundIoInStream *is, int err) {
